requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
langchain-community
langchain-chroma
langgraph
chromadb
pydantic
python-dotenv
